Basic Languages CRUD

+++ src/cat/ehh/web/dao/DAO.java
@@ -1,11 +1,9 @@
 package cat.ehh.web.dao;
 
-import java.math.BigDecimal;
-
 public abstract class DAO<T> {
 
 	public abstract T create(T entity);
-	public abstract T read(BigDecimal entityId);
+	public abstract T read(Long entityId);
 	public abstract T update(T entity);
 	public abstract boolean delete(T entity);
